# schema/openapi.yaml
openapi: 3.1.0
info:
  title: cloudattrib API
  version: "1"
  description: |
    The API accepts at most 1 MiB of JSON request data. The configured operator
    authentication mode applies to every route. A loopback-only deployment uses
    the local-operator identity. Bearer and trusted-proxy modes map each request
    to a stable operator ID. Operator IDs scope idempotency and audit records,
    not result visibility.
servers:
  - url: http://127.0.0.1:8080
security:
  - bearerAuth: []
  - trustedProxyIdentity: []
paths:
  /v1/analyze:
    post:
      summary: Analyze one target synchronously.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./analyze-request.schema.json
      responses:
        "200":
          description: A complete, partial, failed, or cancelled terminal report.
          content:
            application/json:
              schema:
                $ref: ./report.schema.json
        "400":
          $ref: "#/components/responses/InvalidSyntax"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          $ref: "#/components/responses/InputTooLarge"
        "422":
          $ref: "#/components/responses/InvalidTarget"
        "503":
          $ref: "#/components/responses/CapabilityUnavailable"
  /v1/lookup/ip:
    post:
      summary: Look up local IP associations without DNS, HTTP, or persistence.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IPLookupRequest"
      responses:
        "200":
          description: A complete or partial local lookup result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IPLookupResult"
        "400":
          $ref: "#/components/responses/InvalidSyntax"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          $ref: "#/components/responses/InputTooLarge"
        "422":
          $ref: "#/components/responses/InvalidTarget"
        "503":
          $ref: "#/components/responses/CapabilityUnavailable"
  /v1/jobs:
    post:
      summary: Atomically admit a bounded analysis batch.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobSubmission"
      responses:
        "202":
          description: The job was admitted or an idempotent prior admission was returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "409":
          $ref: "#/components/responses/IdempotencyConflict"
        "429":
          $ref: "#/components/responses/CapacityExceeded"
        "503":
          $ref: "#/components/responses/PersistenceUnavailable"
...
  /v1/findings:
    get:
      summary: Search stored findings using deterministic keyset pagination.
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: domain
          in: query
          schema: {type: string}
        - name: provider
          in: query
          schema: {type: string}
        - name: product
          in: query
          schema: {type: string}
        - name: relation
          in: query
          schema: {type: string}
        - name: strength
          in: query
          schema: {type: string}
        - name: observed_from
          in: query
          schema: {type: string, format: date-time}
        - name: observed_to
          in: query
          schema: {type: string, format: date-time}
      responses:
        "200":
          description: Finding page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindingPage"
  /v1/providers:
    get:
      summary: Search the canonical provider catalog.
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/CatalogQuery"
      responses:
        "200":
          description: Provider page.
  /v1/products:
    get:
      summary: Search the canonical product catalog.
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/CatalogQuery"
      responses:
        "200":
          description: Product page.
...
components:
  parameters:
...
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 500, default: 100}
    Cursor:
      name: cursor
      in: query
      schema: {type: string}
    CatalogQuery:
      name: q
      in: query
      schema: {type: string}
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: The configured bearer credential maps to a stable operator ID.
    trustedProxyIdentity:
      type: apiKey
      in: header
      name: X-Cloudattrib-Operator
      description: |
        Use only when the server is configured with a trusted proxy range. The
        proxy removes client-supplied values before it sets this verified header.
  responses:
    InvalidSyntax:
      description: The request body is malformed JSON or contains an unknown field.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: The request did not use the configured trusted-operator identity source.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InputTooLarge:
      description: The request body exceeds 1 MiB.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InvalidTarget:
      description: The request is syntactically valid but has invalid target or option values.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    CapabilityUnavailable:
      description: No necessary execution capability is available for the requested operation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    PersistenceUnavailable:
      description: Durable persistence is unavailable or failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    IdempotencyConflict:
      description: The caller reused an idempotency key for a different request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    CapacityExceeded:
      description: The global nonterminal target backlog is full.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    JobSubmission:
      type: object
      additionalProperties: false
      required: [idempotency_key, targets]
      properties:
        idempotency_key: {type: string}
        bundle_id: {type: string}
        targets:
          type: array
          minItems: 1
          maxItems: 1000
          items:
            $ref: ./analyze-request.schema.json
...
    Job:
      type: object
      required: [id, status, cancel_requested, created_at, updated_at, counts, targets]
      properties:
        id: {type: string}
        status: {type: string, enum: [queued, running, completed, partial, failed, cancelled]}
        bundle_id: {type: string}
        cancel_requested: {type: boolean}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
        counts: {type: object, additionalProperties: {type: integer}}
        targets:
          type: array
          items:
            type: object
            required: [id, input_index, status, attempts]
            properties:
              id: {type: string}
              input_index: {type: integer}
              status: {type: string}
              attempts: {type: integer}
              terminal_reason: {type: string}
              result_url: {type: string}
...
    FindingPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            type: object
            required: [report_id, classified_at, finding]
            properties:
              report_id: {type: string}
              classified_at: {type: string, format: date-time}
              finding:
                $ref: ./finding.schema.json
        next_cursor: {type: string}
    IPLookupRequest:
      type: object
      additionalProperties: false
      required: [address]
      properties:
        address:
          type: string
          format: ip
        match:
          type: string
          enum: [all, longest]
          default: all
        include_retired:
          type: boolean
          default: false
        categories:
          type: array
          items:
            type: string
    IPLookupResult:
      type: object
      required: [address, status, associations, asn, coverage]
      properties:
        address:
          type: string
          format: ip
        status:
          type: string
          enum: [complete, partial]
        associations:
          type: array
          items:
            type: object
        asn:
          type: array
          items:
            type: object
        coverage:
          type: array
          items:
            type: object
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string